# internal/pkg/docs/swagger.yaml
definitions:
  delivery.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.NoteDataForSwagger:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  models.NoteForSwagger:
    properties:
      create_time:
        type: string
      data:
        $ref: '#/definitions/models.NoteDataForSwagger'
      id:
        type: string
      owner_id:
        type: string
      update_time:
        type: string
    type: object
...
  models.UpsertNoteRequestForSwagger:
    properties:
      data:
        $ref: '#/definitions/models.NoteDataForSwagger'
    type: object
...
host: you-note.ru
info:
  contact: {}
  description: API for YouNote service
  title: YouNote API
  version: "1.0"
paths:
...
  /api/note/{id}:
    get:
      description: Get one of notes of current user
      operationId: get-note
      parameters:
      - description: note id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: note
          schema:
            $ref: '#/definitions/models.NoteForSwagger'
        "400":
          description: incorrect id
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
        "404":
          description: note not found
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
      summary: Get one note
      tags:
      - note
...
  /api/note/{id}/delete:
    delete:
      description: Delete selected note of current user
      operationId: delete-note
      parameters:
      - description: note id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: incorrect id
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
        "404":
          description: note not found
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
      summary: Delete note
      tags:
      - note
  /api/note/{id}/edit:
    post:
      consumes:
      - application/json
      description: Create new note to current user
      operationId: update-note
      parameters:
      - description: note id
        in: path
        name: id
        required: true
        type: string
      - description: note data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpsertNoteRequestForSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: note
          schema:
            $ref: '#/definitions/models.NoteForSwagger'
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
      summary: Update note
      tags:
      - note
  /api/note/add:
    post:
      consumes:
      - application/json
      description: Create new note to current user
      operationId: add-note
      parameters:
      - description: note data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpsertNoteRequestForSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: note
          schema:
            $ref: '#/definitions/models.NoteForSwagger'
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
      summary: Add note
      tags:
      - note
  /api/note/get_all:
    get:
      description: Get a list of notes of current user
      operationId: get-all-notes
      parameters:
      - description: notes count
        in: query
        name: count
        type: integer
      - description: notes offset
        in: query
        name: offset
        type: integer
      - description: notes title substring
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: notes
          schema:
            items:
              $ref: '#/definitions/models.NoteForSwagger'
            type: array
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
      summary: Get all notes
      tags:
      - note
...
swagger: "2.0"